airbyte-integrations/connectors/destination-azure-onelake/scripts/requirements-validate.txt
# For validate_onelake_config.py --connectivity (optional)
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
